refactor(orders): wave-4 slice 2 eliminate fixtures_tail and densify shorthand
Remove valid_work_tiles_fixtures_tail, split NW/explore fixtures, add shared
expectation asserts, and shrink the three largest shorthand modules by ≥25%
while ratcheting the support LOC ceiling toward the wave-4 AC.
