+++ packages/colonizethis_orders/test/orders/support/application/work_completion_expectation_shorthand.dart
@@ -1,4 +1,4 @@
-// Compact work-completion expectation shorthands (Refs #3949).
+// Compact work-completion expectation shorthands (Refs #3971).
 
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
@@ -6,6 +6,7 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_orders/src/orders/orders_application_completed_work.dart';
 import 'package:colonizethis_orders/src/orders/orders_application_context.dart';
 import 'package:colonizethis_test/test.dart';
+import '../common/expectation_asserts.dart';
 import 'work_application_fixtures.dart';
 
 Unit wccBuilderImprovement({
@@ -13,17 +14,15 @@ Unit wccBuilderImprovement({
   int totalTurns = 1,
   int remainingTurns = 1,
   bool withOriginAssignment = true,
-}) {
-  return workAppWorkingUnit(
-    type: kUnitTypeBuilder,
-    workTarget: kWorkTargetBuildImprovement,
-    ownerId: ownerId,
-    totalTurns: totalTurns,
-    remainingTurns: remainingTurns,
-    originTileKey: withOriginAssignment ? WorkAppIds.originTileKey : null,
-    assignedTileKey: withOriginAssignment ? WorkAppIds.tileKey : null,
-  );
-}
+}) => workAppWorkingUnit(
+  type: kUnitTypeBuilder,
+  workTarget: kWorkTargetBuildImprovement,
+  ownerId: ownerId,
+  totalTurns: totalTurns,
+  remainingTurns: remainingTurns,
+  originTileKey: withOriginAssignment ? WorkAppIds.originTileKey : null,
+  assignedTileKey: withOriginAssignment ? WorkAppIds.tileKey : null,
+);
 
 Game wccApply(
   Game game, {
@@ -42,241 +41,81 @@ void wccExpectImprovement(
   Game next,
   int level, {
   String tileKey = WorkAppIds.tileKey,
-}) {
-  expect(next.worldState.tileState.improvementLevel(tileKey), level);
-}
+}) => expect(next.worldState.tileState.improvementLevel(tileKey), level);
 
-void wccExpectRoadLevel(Game next, String tileKey, int level) {
-  expect(next.worldState.tileState.roadLevel(tileKey), level);
-}
+void wccExpectRoadLevel(Game next, String tileKey, int level) =>
+    expect(next.worldState.tileState.roadLevel(tileKey), level);
 
 void wccExpectVisibility(
   Game next,
   String tileKey,
   String level, {
   String playerId = 'p1',
-}) {
-  expect(next.worldState.playerVisibilityByTile[playerId]?[tileKey], level);
-}
+}) => expect(next.worldState.playerVisibilityByTile[playerId]?[tileKey], level);
 
-Game wccRailGame({
-  required int roadLevel,
-  List<Player>? players,
-  int turnNumber = 0,
-  bool working = true,
-}) {
+// dart format off
+Game wccRailGame({required int roadLevel, List<Player>? players, int turnNumber = 0, bool working = true}) {
   final unit = working
-      ? workAppWorkingUnit(
-          type: kUnitTypeRailBuilder,
-          workTarget: kWorkTargetBuildRail,
-        )
+      ? workAppWorkingUnit(type: kUnitTypeRailBuilder, workTarget: kWorkTargetBuildRail)
       : workAppUnit(type: kUnitTypeRailBuilder, status: UnitStatus.working);
-  return workAppOwnedGame(
-    turnNumber: turnNumber,
-    units: [unit],
-    tileState: TileMapState().setRoadLevel(WorkAppIds.tileKey, roadLevel),
-    players:
-        players ??
-        [
-          workAppPlayer(techUnlocked: const {kTechIdEarlySteamEngine: true}),
-        ],
-  );
+  return workAppOwnedGame(turnNumber: turnNumber, units: [unit], tileState: TileMapState().setRoadLevel(WorkAppIds.tileKey, roadLevel), players: players ?? [workAppPlayer(techUnlocked: const {kTechIdEarlySteamEngine: true})]);
 }
 
-(BuildWorkState, Unit, CurrentWork) wccDispatchRailSetup({
-  required int roadLevel,
-  List<Player>? players,
-}) {
-  final unit = workAppUnit(
-    type: kUnitTypeRailBuilder,
-    status: UnitStatus.working,
-  );
+(BuildWorkState, Unit, CurrentWork) wccDispatchRailSetup({required int roadLevel, List<Player>? players}) {
+  final unit = workAppUnit(type: kUnitTypeRailBuilder, status: UnitStatus.working);
   final tileState = TileMapState().setRoadLevel(WorkAppIds.tileKey, roadLevel);
-  final game = workAppOwnedGame(
-    turnNumber: 1,
-    units: [unit],
-    tileState: tileState,
-    players:
-        players ??
-        [
-          workAppPlayer(techUnlocked: const {kTechIdEarlySteamEngine: true}),
-        ],
-  );
-  final work = WorkOrderState(
-    unitsById: (oldWorld: {unit.id: unit}, newWorld: const {}),
-    tileState: tileState,
-    visibilityByTile: const {},
-    portsByProvinceSeaboard: const {},
-    purchasedTilesByTileKey: const {},
-    oldProvinces: game.worldState.oldWorld.provinces,
-    newProvinces: const [],
-  );
-  final state = BuildWorkState(
-    game: game,
-    buildOrders: const {},
-    workOrders: const {},
-    tileMapByRegion: {WorkAppIds.ow: workAppRailMap()},
-    work: work,
-  );
-  const cw = CurrentWork(
-    workTarget: kWorkTargetBuildRail,
-    tileKey: WorkAppIds.tileKey,
-    totalTurns: 1,
-    remainingTurns: 0,
-  );
+  final game = workAppOwnedGame(turnNumber: 1, units: [unit], tileState: tileState, players: players ?? [workAppPlayer(techUnlocked: const {kTechIdEarlySteamEngine: true})]);
+  final work = WorkOrderState(unitsById: (oldWorld: {unit.id: unit}, newWorld: const {}), tileState: tileState, visibilityByTile: const {}, portsByProvinceSeaboard: const {}, purchasedTilesByTileKey: const {}, oldProvinces: game.worldState.oldWorld.provinces, newProvinces: const []);
+  final state = BuildWorkState(game: game, buildOrders: const {}, workOrders: const {}, tileMapByRegion: {WorkAppIds.ow: workAppRailMap()}, work: work);
+  const cw = CurrentWork(workTarget: kWorkTargetBuildRail, tileKey: WorkAppIds.tileKey, totalTurns: 1, remainingTurns: 0);
   return (state, unit, cw);
 }
 
-BuildWorkState wccDispatchCompleted(
-  BuildWorkState state,
-  Unit unit,
-  CurrentWork cw, {
-  BuildWorkState Function(BuildWorkState, Unit, String)? onExploreRegion,
-}) => dispatchCompletedWorkTarget(
-  state,
-  unit,
-  cw,
-  () => state.game.worldState.oldWorld.provinces,
-  (w, p) => w.copyWith(oldProvinces: p),
-  onExploreRegion ?? (s, u, regionId) => s,
-);
+BuildWorkState wccDispatchCompleted(BuildWorkState state, Unit unit, CurrentWork cw, {BuildWorkState Function(BuildWorkState, Unit, String)? onExploreRegion}) =>
+    dispatchCompletedWorkTarget(state, unit, cw, () => state.game.worldState.oldWorld.provinces, (w, p) => w.copyWith(oldProvinces: p), onExploreRegion ?? (s, u, regionId) => s);
 
-(BuildWorkState, Unit, CurrentWork) wccDispatchWorkSetup({
-  required Unit unit,
-  required Game game,
-  required CurrentWork cw,
-  Map<String, TileMapResult>? tileMapByRegion,
-  List<Province>? oldProvinces,
-}) {
-  final work = WorkOrderState(
-    unitsById: (oldWorld: {unit.id: unit}, newWorld: const {}),
-    tileState: game.worldState.tileState,
-    visibilityByTile: const {},
-    portsByProvinceSeaboard: const {},
-    purchasedTilesByTileKey: const {},
-    oldProvinces: oldProvinces ?? game.worldState.oldWorld.provinces,
-    newProvinces: const [],
-  );
-  final state = BuildWorkState(
-    game: game,
-    buildOrders: const {},
-    workOrders: const {},
-    tileMapByRegion: tileMapByRegion ?? const {},
-    work: work,
-  );
+(BuildWorkState, Unit, CurrentWork) wccDispatchWorkSetup({required Unit unit, required Game game, required CurrentWork cw, Map<String, TileMapResult>? tileMapByRegion, List<Province>? oldProvinces}) {
+  final work = WorkOrderState(unitsById: (oldWorld: {unit.id: unit}, newWorld: const {}), tileState: game.worldState.tileState, visibilityByTile: const {}, portsByProvinceSeaboard: const {}, purchasedTilesByTileKey: const {}, oldProvinces: oldProvinces ?? game.worldState.oldWorld.provinces, newProvinces: const []);
+  final state = BuildWorkState(game: game, buildOrders: const {}, workOrders: const {}, tileMapByRegion: tileMapByRegion ?? const {}, work: work);
   return (state, unit, cw);
 }
 
-Game wccEngineerCompletionGame({
-  required String workTarget,
-  TileMapState? tileState,
-  List<Province>? provinces,
-  List<Player>? players,
-  Map<String, String>? portsByProvinceSeaboard,
-}) => workAppOwnedGame(
-  units: [workAppWorkingUnit(type: kUnitTypeEngineer, workTarget: workTarget)],
-  tileState: tileState,
-  provinces: provinces,
-  players: players,
-  portsByProvinceSeaboard: portsByProvinceSeaboard,
-);
+Game wccEngineerCompletionGame({required String workTarget, TileMapState? tileState, List<Province>? provinces, List<Player>? players, Map<String, String>? portsByProvinceSeaboard}) =>
+    workAppOwnedGame(units: [workAppWorkingUnit(type: kUnitTypeEngineer, workTarget: workTarget)], tileState: tileState, provinces: provinces, players: players, portsByProvinceSeaboard: portsByProvinceSeaboard);
 
-void wccExpectBuildImprovementCompletesToLevel(
-  int toLevel, {
-  int fromLevel = 0,
-  Map<String, String>? resourceByTileKey,
-  int turnNumber = 0,
-  String ownerId = 'p1',
-  List<Player>? players,
-  Map<String, bool>? aiControlByGpId,
-  String? lastHumanCompletedResearchCategory,
-  int? lastHumanResearchCategoryCompletionTurn,
-}) {
-  final next = wccApply(
-    workAppOwnedGame(
-      turnNumber: turnNumber,
-      units: [wccBuilderImprovement(ownerId: ownerId)],
-      tileState: TileMapState().setImprovement(WorkAppIds.tileKey, fromLevel),
-      resourceByTileKey: resourceByTileKey,
-      provinces: ownerId != 'p1'
-          ? [workAppOwnedProvince(ownerId: ownerId)]
-          : null,
-      players: players,
-      aiControlByGpId: aiControlByGpId,
-      lastHumanCompletedResearchCategory: lastHumanCompletedResearchCategory,
-      lastHumanResearchCategoryCompletionTurn:
-          lastHumanResearchCategoryCompletionTurn,
-    ),
-  );
+void wccExpectBuildImprovementCompletesToLevel(int toLevel, {int fromLevel = 0, Map<String, String>? resourceByTileKey, int turnNumber = 0, String ownerId = 'p1', List<Player>? players, Map<String, bool>? aiControlByGpId, String? lastHumanCompletedResearchCategory, int? lastHumanResearchCategoryCompletionTurn}) {
+  final next = wccApply(workAppOwnedGame(
+    turnNumber: turnNumber,
+    units: [wccBuilderImprovement(ownerId: ownerId)],
+    tileState: TileMapState().setImprovement(WorkAppIds.tileKey, fromLevel),
+    resourceByTileKey: resourceByTileKey,
+    provinces: ownerId != 'p1' ? [workAppOwnedProvince(ownerId: ownerId)] : null,
+    players: players,
+    aiControlByGpId: aiControlByGpId,
+    lastHumanCompletedResearchCategory: lastHumanCompletedResearchCategory,
+    lastHumanResearchCategoryCompletionTurn: lastHumanResearchCategoryCompletionTurn,
+  ));
   wccExpectImprovement(next, toLevel);
-  final after = wccSingleUnit(next);
-  expect(after.tileKey, WorkAppIds.tileKey);
-  expect(after.originTileKey, isNull);
-  expect(after.assignedTileKey, isNull);
+  expectUnitIdleCleared(wccSingleUnit(next), tileKey: WorkAppIds.tileKey);
 }
 
-Game wccExploreWorkingGame({
-  Map<String, Map<String, List<String>>>? tileKeysByRegionAndProvince,
-  Map<String, Map<String, String>>? playerVisibilityByTile,
-}) => workAppOwnedGame(
-  units: [
-    workAppWorkingUnit(
-      type: kUnitTypeExplorer,
-      workTarget: kWorkTargetExplore,
-    ),
-  ],
-  tileKeysByRegionAndProvince:
-      tileKeysByRegionAndProvince ??
-      {
-        WorkAppIds.ow: {
-          WorkAppIds.provinceId: [WorkAppIds.tileKey],
-        },
-      },
-  playerVisibilityByTile: playerVisibilityByTile,
-);
+Game wccExploreWorkingGame({Map<String, Map<String, List<String>>>? tileKeysByRegionAndProvince, Map<String, Map<String, String>>? playerVisibilityByTile}) =>
+    workAppOwnedGame(units: [workAppWorkingUnit(type: kUnitTypeExplorer, workTarget: kWorkTargetExplore)], tileKeysByRegionAndProvince: tileKeysByRegionAndProvince ?? {WorkAppIds.ow: {WorkAppIds.provinceId: [WorkAppIds.tileKey]}}, playerVisibilityByTile: playerVisibilityByTile);
 
-void wccExpectFullyVisible(Game next, String tileKey) =>
-    wccExpectVisibility(next, tileKey, VisibilityLevel.fullyVisible.name);
+void wccExpectFullyVisible(Game next, String tileKey) => wccExpectVisibility(next, tileKey, VisibilityLevel.fullyVisible.name);
 
 void wccExpectRailApply(int roadLevel, int expected) {
-  final next = wccApply(
-    wccRailGame(roadLevel: roadLevel),
-    tileMapByRegion: {WorkAppIds.ow: workAppRailMap()},
-  );
+  final next = wccApply(wccRailGame(roadLevel: roadLevel), tileMapByRegion: {WorkAppIds.ow: workAppRailMap()});
   wccExpectRoadLevel(next, WorkAppIds.tileKey, expected);
 }
 
-void wccExpectRailDispatch(
-  int roadLevel,
-  int expected, {
-  List<Player>? players,
-}) {
-  final (state, unit, cw) = wccDispatchRailSetup(
-    roadLevel: roadLevel,
-    players: players,
-  );
-  expect(
-    wccDispatchCompleted(state, unit, cw).work.tileState.roadLevel(
-      WorkAppIds.tileKey,
-    ),
-    expected,
-  );
+void wccExpectRailDispatch(int roadLevel, int expected, {List<Player>? players}) {
+  final (state, unit, cw) = wccDispatchRailSetup(roadLevel: roadLevel, players: players);
+  expect(wccDispatchCompleted(state, unit, cw).work.tileState.roadLevel(WorkAppIds.tileKey), expected);
 }
 
-Game wccRoadPropagateGame({
-  required int workTileRoad,
-  required int adjacentRoad,
-  required String adjacentTileKey,
-  List<Player>? players,
-  Map<String, String>? portsByProvinceSeaboard,
-}) => wccEngineerCompletionGame(
-  workTarget: kWorkTargetBuildRoad,
-  tileState: TileMapState()
-      .setRoadLevel(WorkAppIds.tileKey, workTileRoad)
-      .setRoadLevel(adjacentTileKey, adjacentRoad),
-  players: players,
-  portsByProvinceSeaboard: portsByProvinceSeaboard,
-);
+Game wccRoadPropagateGame({required int workTileRoad, required int adjacentRoad, required String adjacentTileKey, List<Player>? players, Map<String, String>? portsByProvinceSeaboard}) =>
+    wccEngineerCompletionGame(workTarget: kWorkTargetBuildRoad, tileState: TileMapState().setRoadLevel(WorkAppIds.tileKey, workTileRoad).setRoadLevel(adjacentTileKey, adjacentRoad), players: players, portsByProvinceSeaboard: portsByProvinceSeaboard);
 
 void wccExpectCancelledIdleAtOrigin(Game next) {
   final u = wccSingleUnit(next);
@@ -287,26 +126,9 @@ void wccExpectCancelledIdleAtOrigin(Game next) {
   expect(u.assignedTileKey, isNull);
 }
 
-(BuildWorkState, Unit, CurrentWork) wccDispatchTargetSetup({
-  required String unitType,
-  required String workTarget,
-  List<Province>? provinces,
-}) {
+(BuildWorkState, Unit, CurrentWork) wccDispatchTargetSetup({required String unitType, required String workTarget, List<Province>? provinces}) {
   final unit = workAppUnit(type: unitType, status: UnitStatus.working);
-  final game = workAppOwnedGame(
-    turnNumber: 1,
-    units: [unit],
-    provinces: provinces ?? const [],
-  );
-  return wccDispatchWorkSetup(
-    unit: unit,
-    game: game,
-    cw: CurrentWork(
-      workTarget: workTarget,
-      tileKey: WorkAppIds.tileKey,
-      totalTurns: 1,
-      remainingTurns: 0,
-    ),
-    oldProvinces: provinces ?? const [],
-  );
+  final game = workAppOwnedGame(turnNumber: 1, units: [unit], provinces: provinces ?? const []);
+  return wccDispatchWorkSetup(unit: unit, game: game, cw: CurrentWork(workTarget: workTarget, tileKey: WorkAppIds.tileKey, totalTurns: 1, remainingTurns: 0), oldProvinces: provinces ?? const []);
 }
+// dart format on
